Move layer features formats into separate folder
